+++ CoreOS/system/kernel/purr_version.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#define PURR_OS_VERSION       "0.9.4"
+#define PURR_OS_VERSION       "0.9.5"
 #define PURR_OS_VERSION_MAJOR 0
 #define PURR_OS_VERSION_MINOR 9
-#define PURR_OS_VERSION_PATCH 4
+#define PURR_OS_VERSION_PATCH 5
 
-#define KITT_VERSION          "0.5.3"
+#define KITT_VERSION          "0.6.0"
 #define KITT_VERSION_MAJOR    0
-#define KITT_VERSION_MINOR    5
-#define KITT_VERSION_PATCH    3
+#define KITT_VERSION_MINOR    6
+#define KITT_VERSION_PATCH    0